feat(CI/TEST): split ci/cd between dev and main + remove test from build
