+++ src/index.js
@@ -21,7 +21,7 @@ app.set('port', port);
 mongoose.set('strictQuery', false);
 
 mongoose.connect(
-  `mongodb+srv://${process.env.MONGO_DB_USER}:${process.env.MONGO_DB_PASSWORD}${process.env.MONGO_URL}/?retryWrites=true&w=majority`,
+  `mongodb+srv://${process.env.MONGO_DB_USER}:${process.env.MONGO_DB_PASSWORD}$${process.env.MONGO_URL}/?retryWrites=true&w=majority`,
 ).then(() => {
   console.log('successfully connect to database');
 }).catch((err) => console.log(err));
